Even if multiple tokens match in a particularly entry, only add it to an entryBucket once
Oh man, this one was a doozy. Before this change it was possible for 'The Titan' to add two entries to the 't' bucket. After deduping you'd still only see a single entry, but the problem was that the count would end up being wrong. Also tons of wasted memory storing redundant entries.

This should basically fix all bugs observed in the movies example. I still need to profile this though because I'm a bit worried about all those calls to Array::includes.