Avoid doing unnecessary extra work when checking if tokens match
While debugging another problem (the 'allow-matching-the-same-token-multiple-times' problem) I noticed that we're incorrectly only incrementing numEntryTokensChecked in the DataSource when we don't have a matching token. This means we end up just doing an extra iteration over the while loop. Instead, increment always, even when there's a match, since one of the entryTokens was obviously checked when there's a match.