Add support for keywords to DataSource!
I landed the previous commits to make this as focused on the actual keyword functionality as possible. Basically, we append entries into entryBuckets based on their keywords after we've processed entries based on their actual display text. This is a pretty simple implementation and for example, you can't search for multiple keywords out of order, because of the existing search rules / behavior, but I'm pretty happy with the simplicity of the implementation.