Do proper accent and diacritic cleaning using the diacritic dependency
I don't feel great about this dependency since it's 7 years old and probably overkill for what we need, but I don't want to create a modern alternative just yet. I think we should be able to greatly simplify this with String.prototype.normalize and a regular expression, but I'm worried that will be even slower than this. For posterity, it looks like this slows down parseTokens by 2-3x :( but it's a one-time cost and this should only matter with DataSource instances that have ~100K entries or more. This is good enough for now.