Store selected entry in typeahead when one is chosen
This enables us to style the input based on whether a value is selected, and provide visual feedback when the entry is no longer chosen. This also changes onReset to be called whenever the input value changes after a selection, which is the right behavior. I may want to change the name of onReset in the future, but this works for now.