Refactor pages directory and move things around
I wanted to play around with dynamic routes and dynamic imports, so I figured I'd change the way things are structured a bit. This moves all the 'pages/example.js' to 'src/examples/example/index.js' to collocate them with their other resources. It also removes the need for each example to have to import the Template, and simplifies that redirection.

I've given up the custom page titles for each example because they're not necessary, but they can be added back with next/head and the title component.