Implement suggested text behavior
This implements a hidden input sitting right over the top of the main Typeahead input to provide the 'suggested text' bhavior. If the flattened highlighted result's text starts with the flattened input query, the rest of the match will be displayed in the input, in its original format. This builds on quite a few other diffs, like splitting out flatten, some Typeahead refactoring, etc.