Stop passing the Typeahead's text input to the onSelect callback
This is accomplished by using the return value from invoking onSelect as the text input's value, if one is provided. Otherwise, the default and most common behavior will be for the input's value to be set to entry.getText(). This also required getting rid of the clipboard hack and using the proper navigator.clipboard.writeText in the emoji example to retain that functionality.