Create proper dist script
The exported (to npm) typeahead package doesn't depend on either next or emojibase, only the examples depend on those things. Unfortunately, I can't 'have my cake and eat it too'. I want the next project to be super simple, and not a workspaced monorepo, so that changes to the core files will fast-refresh the examples. In order to make this happen I have to actually create a separate 'build process' that leverages a completely different package.json. This seems to work, will be doing some testing in another project now.