Fix backspace no longer working
When I updated this for the movielist example, to make it so that a DataSource response wouldn't be displayed after a selection had been made, I failed to consider the backspace case. This is a stop-gap fix that only actually ignores the response if the input value is empty when an actual list of results is returned.

In the future, if I want to have 'default entries' that show up on focus, this won't work, but it's good enough for now.